ascii/requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
Pillow==7.1.2
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.11.2
